refactor config and run to new file
